==> analysis/02-run-regions.R <==
source("analysis/model-prep.R")

source(file.path(this_folder, "analysis", "01-fit-british-columbia.R"))
source(file.path(this_folder, "analysis", "01-fit-ontario.R"))
source(file.path(this_folder, "analysis", "01-fit-quebec.R"))

source(file.path(this_folder, "analysis", "01-fit-california.R"))
source(file.path(this_folder, "analysis", "01-fit-washington.R"))
source(file.path(this_folder, "analysis", "01-fit-new-york.R"))

source(file.path(this_folder, "analysis", "01-fit-belgium.R"))
source(file.path(this_folder, "analysis", "01-fit-new-zealand.R"))
source(file.path(this_folder, "analysis", "01-fit-sweden.R"))
source(file.path(this_folder, "analysis", "01-fit-uk.R"))
source(file.path(this_folder, "analysis", "01-fit-japan.R"))
source(file.path(this_folder, "analysis", "01-fit-germany.R"))
